Split models by concern.
